power: Log suspend_stress_test completion time
The script suspend_stress_test typically runs for many hours unattended,
so it's not always known exactly when it encountered an error. This CL
outputs the time in UTC when the script completes, either successfully
or due to an error, to make it easier to correlate with the device logs.

BUG=None
TEST=suspend_stress_test, force premature wake, verify time is output
  Suspend     1 of 1875: sleep for 18 seconds...�Premature wake detected: slept for 3s, expected 18s, test failed.
   (dmesg > /usr/local/agingLogs/dmesg_1)...
  Test completed at: Wed May 31 17:52:39 UTC 2023

Change-Id: I6847460111034565ffa27fb82820c204d961f60c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4577402
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Tim Van Patten <timvp@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>