hiberman: Report hibernate abort metric when 'hiberimage' is not set up
Report a hibernate abort metric when 'hiberimage' is not set up
when hiberman is asked to hibernate.

'hiberimage' is set up on the first login after boot, and torn
down when the user logs out. It is currently not set up again
on subsequent logins. The new metric value can help with
understanding how many users are impacted by this limitation.

BUG=b:280452479, b:232184089
TEST=hiberman hibernate -r
  hiberman teardown-hiberimage
  hiberman hibernate
  chrome://histograms has value 4 in Platform.Hibernate.Abort

Change-Id: I50cd462923f1cb2b420cfaa954c6e3b388cb201b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4572775
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>