system_api: add charge_limited to power_supply_properties.proto
This adds a bool to the PowerSupplyProperties proto that's sent to
Chrome to indicate whether charging is limited due to the Charge Limit
feature. This bool will be set to true if the battery percentage is
close to or above the battery percentage limit.

BUG=b:280880357
TEST=emerge-atlas chromeos-base/system_api

Change-Id: Ia34677258860621495f7b7a29202db3605a0d1cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4576517
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>