cryptohome: Add some enum<->enum conversions
This adds library functions to convert between our internal enum types
and the new matching flatbuffer enums.

This also fixes some problems with the original flatbuffer codegen which
it seems was wrong. Unfortunately due to a typo the code wasn't being
generated the way I thought, and this only became apparent when I tried
to actually make use of the headers in code. Unfortunately it turns out
that yup, we definitely need to move the shared .fbs off to a separate
namespace so that we can filter it out of the codegen for the other leaf
.fbs files.

BUG=b:281879458
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib379dea6f3eee0a97f63f29e067925d62ea75af5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4581989
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>