cryptohome: Move protobuf conversion functions out of utils
This is just an opportunistic change from when I was creating the
similar flatbuffer conversion functions. In general "utils" dumping
grounds are not great, and it's pretty easy to just put the protobuf
specific stuff here in a separate protobuf module. Plus it adds a bit of
symmetry with the flatbuffer one.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ibbfb1bb24b5f86a40e6d0a55a477730cb8d24a6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4581630
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>