secanomalyd: adds new filter for proc enumeration
Adds a new filter to the process enumeration mechanism for excluding the
kernel tasks. This is done by parsing the PPID field from the status
file and excluding processes that have a parent PID of 2, which is the
kthreadd PID. In addition processes that have a PPID of 0 are also
excluded, which includes the kthreadd process itself and the init
process.
Updates process enumeration tests to include the new PPID attribute and
test the parsing of the new attribute, as well as the filtering
mechanism.
Updates reporter tests to include PPID attribute for ProcEntry objects.

BUG=b:287087910
TEST=Manually verified processes are being filtered correctly, by
issuing a crash report with each type of filter.
TEST=All unit tests pass.

Change-Id: I7629388375afa4c91bb7b0567ba795ca08789a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4611070
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>