secanomalyd: adds seccomp coverage metric
Adds a metric for reporting the SecComp coverage on the system. This is
done by dividing the number of processes covered by SecComp by the total
number of processes on the system and the metric is sent as a
percentage. The metric is sent once during the lifetime of the daemon,
and it is sent only when the user is logged in. In case of a failure, it
is retried every two hours.
To facilitate refreshing the login state without updating the state of
known mount observations, a new argument is introduced for the Refresh
function to selectively limit refreshing only to the system login state.

BUG=b:275757413
TEST=The metric is emitted only once and only if logged in.
TEST=Manually verified the calculated percentage.
TEST=All unit tests pass.

Change-Id: I8bcfd3162aa62d2884beceb1c8041d9c97a95f85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4581992
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>