shill: cellular: Ensure service properties are stored
If a property is not set via dbus, or the property is modified and a
connection is never made, the property will not be saved in the cache.
This is not a problem for other technologies because the profile is
saved when the Manager calls Stop. For cellular, the provider destroys
all the services before the Manager calls Stop, so any values modified
after the last Save, will not be stored.
Ensure all values are stored by calling PersistService before
destroying the cellular service.

BUG=b:233910184
TEST=Trigger an unsuccessful connect and verify that the property is
stored when shill is stopped.

Change-Id: Id4d82b073e6a1d22c44412647d0087ef9348b4bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4581636
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>