printscanmgr: remove StopUpstartJob method
Now that printscanmgr uses lorgnette's D-Bus API to control the debug
flag for lorgnette, there are no more calls to StopUpstartJob. We don't
want to allow the root-level executor daemon any more functionality than
necessary, so we should remove this method.

Note that the debug logging method should eventually control
ippusb_bridge, which it doesn't today, so this functionality may be
needed again. However, if it is, it will be easy to add back.

BUG=b:257070388
TEST=FEATURES=test emerge-nami printscanmgr

Change-Id: Ie9c8e9a6fe079eba751eca7a1bb5548827e799e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4609032
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>