shill: wifi: Report more accurate signal strength
This change intends to make sure that the WiFi quality indicator in the
UI doesn't jump drastically from indicating no signal to an excellent
signal as the result of a single bad packet.

The first new quality indicator is the average signal strength.  This
value will change more slowly and be less affected by a single low
signal quality packet.

If the average signal strength is not available, the preferred quality
indicator will be the average beacon signal strength.

If neither of the average beacon signal or average signal are available,
the code will behave as it did previously.

BUG=b:274971131
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: Ibfdeb38798730d7ccc22edcecf6f9335b76fd544
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4595516
Tested-by: David Ruth <druth@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: David Ruth <druth@chromium.org>