diagnostics: Read part of uploads.log for crash
Currently the whole uploads.log is read for an uploaded crash. Change
this behavior to the more efficient reading only the necessary part of
uploads.log.

BUG=b:266018436
TEST=FEATURES=test emerge-$BOARD diagnostics
TEST=Install the modified binaries to a DUT. Follow the instructions in
the comments of crash_events.h to monitor crash events. Inspect that
both uploaded and unuploaded crashes are emitted to the subscriber.

Change-Id: Ia36e4c247a2cdf8bf58a8e36d0a371ac1dc1fb18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4585902
Tested-by: Hong Xu <xuhong@google.com>
Commit-Queue: Hong Xu <xuhong@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>