patchpanel: rename DBus method to GetDownstreamNetworkInfo
Originally, one of the patchpanel's DBus method name is
DownstreamNetworkInfo. But in convention the naming of the DBus method
should start from a verb.

This CL changes it to "GetDownstreamNetworkInfo" to reflect the
intention. Also, the protobuf message for its argument and response
are renamed to "GetDownstreamNetworkInfoRequest" and
"GetDownstreamNetworkInfoResponse".

BUG=b:273750922
TEST=FEATURES=test emerge-brya chromeos-base/system_api patchpanel \
     patchpanel-client dns-proxy

Change-Id: Ie741260e7c1fc3ccf2877744e5110a0db0847292
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4583778
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>