dlp: Add DlpAdaptor::AddFiles
Add a new API to add to the daemon several files at once to minimize the
number of DBUS calls in case more than once file has to be added.
The idea is to probably deprecate the existing AddFile in place of this
new AddFiles.

BUG=b:285868839
TEST=cros_workon_make --board=${BOARD} chromeos-base/dlp --test

Change-Id: I357cb96e58dd9e6ce7fe10400875ad9a73bfcf60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4593717
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Tested-by: Luca Accorsi <accorsi@google.com>
Commit-Queue: Luca Accorsi <accorsi@google.com>