login: move killers logic here
We've dropped all other users of the killers code in the tree, so move
the helper script to the one last project that relies on it.  This will
make it a bit more obvious when we can finally delete this user of it
too that we can drop the killers logic entirely.

BUG=b:286078477
TEST=CQ passes

Cq-Depend: chromium:4595907
Change-Id: I76ca660aefcdccce2b4243c29e8c09f21556c879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4594520
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>