vm_tools: refvm: Upload built image in upload_image step
Build a brotli-compressed qcow2 to ensure sparse storage and transparent
decompression by the installer.

The image must be uploaded with the Content-Encoding header set, which
isn't supported by kokoro's builtin uploader. This build step can be
extended to create a CL to update tast-tests with the new image and the
hash of the uncompressed image.

BUG=b:279686117
TEST=kokoro

Change-Id: I7d1e3af9e30ebfc222613b782679891a03e8ea2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4600827
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: James Ye <jamesye@google.com>
Reviewed-by: David Munro <davidmunro@google.com>