secanomalyd: adds non-root process percentage UMA
Adds a metric for reporting the percentage of processes that are
not root. This is done by parsing the real UID for each process,
while reading the status file and setting the non-root bit in the
sandboxing bitmap of the process entry if the UID is not 0.
Adds the non-root bit to the process tests so that the parsing is
validated.

BUG=b:285949894
TEST=All tests pass.
TEST=Manually confirmed the number of root processes with the reported
number.
TEST=The UMA metric emits only once with the correct percentage.

Change-Id: I837743920f56a6e6038f5d2d03fe3ec7080c3c31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4599783
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>