vm_tools: Fix possible out-of-bounds read in notificationd
Theoretically if a button_index larger than the number of buttons were
sent to notificationd from exo we'd read past the end of the array. It's
not performance-sensitive code so let's make it a CHECK instead of a
DCHECK to always crash if that happens.
Also fix a typo while I'm at it.

BUG=b:286204915
TEST=CQ, #yolo

Change-Id: Ic2aa5d7c01ca85b99cc59ca20ba4e1630ac03c49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4600416
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: David Munro <davidmunro@google.com>
Tested-by: David Munro <davidmunro@google.com>