u2fd: Only use SN in hex for KeyId
Currently we use the whole DER encoding of the SN field for KeyId's
data. Because CryptAuth only supports reading the SN field as an int and
encoding it back to hex string, it can't produce the DER encoding. The
DER encoding without the first 2 bytes should be equivalent to what they
produce, under the condition that SN is always positive. This currently
holds for cr50 and ti50.

BUG=b:241797804
TEST=u2fd-client unittests

Cq-Depend: chromium:4623527
Change-Id: Ia2459f4f055fe269db5488aac24c9071a70ec10a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4623587
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>