arc: Use arcpp-media-sharing-services for ARC P external storage
The new Upstart job triggers arc-myfiles* and arc-removable-media* in a
way similar to arcvm-media-sharing-services.

At the moment, it's just used by ARC P container, and there is no
visible addition to the functionality of the current implementation.

In ARC R container, it will be started by Chrome and propagate the UID
of Android's MediaProvider to arc-myfiles*. The UID will be used to
calculate the SELinux context for "My files" on the ARC side.

With the planned changes for ARC R container, this will replace the
external storage sharing logic introduced by CL:4390706.

BUG=b:286238422
TEST=arc.MyFiles, arc.RemovableMedia
TEST=initctl list | grep -e myfiles -e removable-media

Cq-Depend: chromium:4604330
Change-Id: I8d3b298b7d692ce2723ec1cd45c1f98dea578232
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4603692
Reviewed-by: Yury Khmel <khmel@google.com>
Reviewed-by: Momoko Hattori <momohatt@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>