secagentd: rework network bpf garbage collection
Previously garbage collection was done by maintaining a list
of affected tupltt es in a map. When a socket was released it would then
mark all associated tuples as to-be-garbage-collected.
In the case where  the number of tuples exceeded the capacity of the
list the flow map would be iterated through and all associated tuples
would be marked for garbage collection.

Unfortunately in 5.10 kernels the map iteration helper isn't available.

so now, we create a map and just insert released socket ids into that
map. Userspace will iterate through that released socket map and then
garbage collect relevant entries on its own.

The increased computational complexity is anticipated to be small since
relatively few sockets should be released over a period of time.

BUG=b:264550046
TEST=manually tested and verified that things are being garbage
collected.

Change-Id: I36a187a85381e65e35284b32e138ba9735eb6bc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4603715
Reviewed-by: Ryan Borzello <rborzello@google.com>
Commit-Queue: Jason Ling <jasonling@chromium.org>
Tested-by: Jason Ling <jasonling@chromium.org>
Auto-Submit: Jason Ling <jasonling@chromium.org>