secanomalyd: adds unpriv proc percentage UMA
Adds a new metric for reporting the percentage of processes that do not
have CAP_SYS_ADMIN and are not root. A new sandbox bit is added to the
existing ProcEntry data structure that tracks whether the process has
the CAP_SYS_ADMIN capability. This is set by parsing the effective
capabilities field in the process status file.

BUG=b:285949894
TEST=The UMA histogram emits only once when logged in and shows the
correct value.
TEST=All tests pass.

Change-Id: Ie4a78ba7e90b8fa1ba0d31cd568a394681a8f430
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4605743
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>