sepolicy: Add proc_timer label to correct genfs_contexts
To label to the timer_highres proc node, we previously modified the base
genfs_contexts instead of chromeos. Analyzing the policy.30 file that is
generated shows the genfscon rule does not make it when doing so. Doing
it in the chromeos genfs_contexts instead works. So fix it.

Also use the correct path to the proc node.

With this I see the files on the device get correctly updated:

/etc/selinux/arc/contexts/files/file_contexts:/proc/sys/kernel/timer_highres         u:object_r:proc_timer:s0
grep: /etc/selinux/arc/policy/policy.30: binary file matches

TEST=analyze the policy.30 file on device.
BUG=b:263289152

Change-Id: Ie920268902455c2ae598fd3710ad34d10ee57db8
Signed-off-by: Joel Fernandes <joelaf@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4606196
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>