power: Don't report Power.TimeInSuspendAt{Boot,Resume} for hibernations
Reporting Power.TimeInSuspendAt{Boot,Resume} for hibernations would
be misleading, especially for the 'Boot' metric, which is recorded
when a device runs out of battery during suspend. Don't write the
suspend timestamp to 'suspend-to-ram-time' for hibernations, which
causes 'send_metrics_on_resume' to skip reporting the metrics.

BUG=b:232184089
TEST=powerd_dbus_suspend --suspend_for_sec=8
    => Power.TimeInSuspendAtResume is reported
  hiberman hibernate -r
    => Power.TimeInSuspendAtResume is not reported

Change-Id: I8eca4a8117a7f157c7aeb18513dc80283785edde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4629152
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>