common-mk: Suppress new warning on C++20 switching.
This is preparation to enable C++20 build option.
Some new compile warning causing build error. Supressing the warning
for the short term to move forward, as this is not a regression.

We should be able to work on actual fixes separately.

BUG=b:217226507
TEST=Tryjob

Change-Id: I25a76dbb7f2f8f8b1ce83dbf842cc0da6efbc69d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4607742
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Eric Lok <lokeric@google.com>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>