shill: device: avoid checking fixed_ip_params() setting on interface down
The fixed_ip_params is used by the ARC VPN to specify that the Network
in use is not managed by shill.

BringNetworkInterfaceDown() should not be called by the ARC VPN device
and it should anyway be up to the device judgement whether to bring up
or down the interface, not the Network.

The BringNetworkInterfaceDown() method can now be called
unconditionally on the device interface_index() regardless of whether
a Network exists or not.

BUG=b:265607618
TEST=cros_workon_make --board=$BOARD --test shill

Change-Id: I5858929fd1110a5b1380e815b56d28e98e2799c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4608581
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>