attestation: Move context out of SignEnterpriseChallenge
The attestation_service is now relying on the input from chromium
instead of deciding itself how to construct the VA challenge response.
This means that it is passing the domain and flow_type directly through
the challenge response and uses the flag `include_certificate` instead
of deciding implicitly based on whether the challenge is signed by a
user or device key. This will make the creation of the VA challenge
responses more extensible with respect to adding more VA flow types.
For more context, see go/remote-attestation-unmanaged-devices-dd.

This must land after chrome-in-ChromeOS has been upreved to a chrome
revision that contains
https://chromium-review.googlesource.com/c/chromium/src/+/4609122 which
implements the relevant logic in Chrome.

BUG=b:277709127
TEST=tast run <ip> hwsec.Attestation* (on chromium:4614864)

Cq-Depend: chromium:4614864
Change-Id: I0ae01e16443c49b4cfcaedc82fddf07070b0edf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4614803
Commit-Queue: Leon Masopust <lmasopust@google.com>
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Leon Masopust <lmasopust@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>