hps: Reset HPS if camera times out
We are seeing some instances of hpsd crashing when the camera fails to
produce an image on time. Instead of crashing the entire daemon,
power-cycle HPS instead to try and recover from this fault. This should
restore HPS functionality faster (~8s) compared to relying on upstart
restarting the daemon.

BUG=b:266351818
TEST=unit tests

Change-Id: I9521eaeb9b72592ff44d3a7da1ca6e40949072f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4603875
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Tested-by: Sami Kyöstilä <skyostil@chromium.org>