croslog: small readability/input check improvement on ParseTime
This addresses the concern that we are reading OOB, although
there is a check in ParseInternal to make sure the
input size is large enough. Still, moving this size check
to the branch condition would make it more readable,
as it becomes a typical size check before read.

BUG=b:286984320
TEST=Unit tests

Change-Id: I9cba411803504ceb5ed05ab40cfd2cbced21de97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4611067
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Zi Lin <lziest@google.com>
Tested-by: Zi Lin <lziest@google.com>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>