shill: device: only run IP config update if connected
Running the DHCP renew and IPv6 config invalidation procedures makes
only sense when the device is connected, so explicitly check for that.

This change also helps integrate the new Network management logic in
Cellular, because one of the assumptions done is that a Network
object is always available if the Device is connected.

BUG=b:265607618
TEST=cros_workon_make --board=$BOARD --test shill

Change-Id: Ifb29b7cab84fe4abbba7bae05bfbee2b865ca691
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4610205
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>