spaced: Add madvise syscall
At present, spaced cannot use the madvise syscall.
With the glibc allocator, this seems to be fine.

However, other allocators such as the LLVM scudo
hardened allocator use madvise to hint to the kernel
that a given address range can be freed up.

Add this syscall so that we don't hit a seccomp with scudo.
madvise is a fairly safe syscall, but even so,
we can restrict it to only MADV_DONTNEED for now.

BUG=b:248085566
TEST=Tested on crosfleet volteer-scudo board

Change-Id: I5b0e72e69419aef909eecce7fc60b50c6c29a415
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4611075
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>