shill: cellular: Move the logic to skip attach round robin
Move the logic to skip the round robin for attach APNs into
BuildApnTryList so all the APN list logic is in one place.
This logic is only used when using the old APN UI, so add that
condition in the if condition.
Remove the NULL APN since the fallback APN already covers the same use
case.
This also moves the logic that adds the first of multiple Attach
APNs as a fallback to BuildAttachApnTryList.

BUG=None
TEST=tast run crota cellular.ShillCustomApn.*
tast run limozeen cellular.ShillCustomApn.*

Change-Id: I27b68de2a4e86cc63b73413d60e170c377d39374
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4622273
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>