cryptohome: Support auth_intents_for_types
This adds support to ListAuthFactors for filling out this new field. The
initial implementation here just reports current==min==max for the
intents for each type because we don't yet have any support for having
factors which are supported-but-disabled (or supported-by-not-enabled).
Note that this doesn't mean the reporting is inaccurate, just that the
distinction is not yet useful because the thing being distinction does
not yet exist.

BUG=b:281878872
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I9a7312efe07b971abccff0428cdcb68274b5727d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4616426
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>