missive: Return error status if storage pointer not valid.
When StorageModule switches implementations of storage, it must drop its
reference to storage before creating a new storage. This temporarily
invalidates the pointer to storage and renders StorageModule unavailable
to handle incoming requests, so during this period we must return an
error to callers.

BUG=b:286967362, b:237809917
TEST=FEATURES=test emerge-${BOARD} missive

Change-Id: I57bb457a1a1c24dd8806eb94d70e317ebbf6ed9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4615835
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Josh Hilke <jrhilke@google.com>