shill: device: only process events from the primary network
In non-Cellular devices, the primary Network will be associated to the
same interface as the Device. In Cellular devices, the primary Network
may be a multiplexed network interface with a different index than the
Device interface index.

Only the events happening in the primary Network are processed by
the device, any other event in different Networks are silently
dropped.

BUG=b:249386791
TEST=cros_workon_make --board=$BOARD --test shill

Change-Id: If48e54b967dcc06f2637793f66abf3c338686afa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4614689
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>