patchpanel: fix net_base type creation at fuzzer test
This CL fixs the crash when creating net_base::IPv4Address and
IPv6Address at fuzzer test. The root cause of the crash is that
FuzzedDataProvider::ConsumeBytes(bytes) might return a vector with
the smaller size. Therefore the IP address creation might fail due to
that. This CL fallback this case to default value.

BUG=b:287264391
TEST=USE="$USE asan fuzzer" FEATURES=test emerge-brya \
     patchpanel patchpanel-client

Change-Id: I6f2268d535bcbfadc7dce8622f3650fd0e18817f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4617310
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>