cryptohome: Add infrastructure for current/min/max intents
This extends the driver library to add in the necessary bits for making
intents configurable. As of this change you can't actually do the
configuration (need the backing policy file) the API support is in place
and the drivers now have a concept of always-on vs on-by-default vs
off-by-default vs always-off.

BUG=b:281878872
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I063492ed27e39be6b8140760763529b11da4eede
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4618799
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>