power: Change PowerSupply to take TimeDelta instead
This changes the function SetAdaptiveCharging for the
PowerSupplyInterface to take a TimeDelta instead of TimeTicks as a
parameter. The PowerSupply only cares about the TimeDelta that it
calculates from the TimeTicks value, and we had a clock mismatch bug.
This simplifies things and fixes that bug.

BUG=b:287558776
TEST=unittests, tast.power.AdaptiveCharging

Change-Id: I9a55f49c73b45c0b3c90479f3cc2c8f4dda92443
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4619867
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>