camera: Fix race condition in effects still capture by copying
The YUV buffer for still capture should not be returned to the client
and sent to still capture processor at the same time, since the buffer
handle may be freed. Fix this issue by making a copy for still capture.

This is a different approach to the previous fix crrev.com/c/4592632
which caused problem in returning out-of-order capture results.

BUG=b:281494965, b:287130947
TEST=Take pictures in CCA.

Change-Id: I400f9e19ba9e615b7592aa4628d543245fd30e4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4617794
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Ren-Pei Zeng <kamesan@chromium.org>