login: Set root_disk_serial_number before we dump the VPD
This prevents any possibility that someone overrides its value through
the VPD, which would provide an FRE escape for some devices.

In addition to the unit test that checks that we record the first value
that we see for a given key, we did some manual testing as follows on a
live system on a DUT:

  % vpd -i RW_VPD -s root_disk_serial_number=BAD_VALUE
  % /usr/share/cros/init/ui-collect-machine-info
  % grep root_disk_serial_number /run/session_manager/machine-info

seeing that root_disk_serial_number="BAD_VALUE" is indeed the second
occurrence of the root_disk_serial_number key in the machine-info file.

BUG=b:288000192
TEST=FEATURES=test emerge-${BOARD} chromeos-login

Change-Id: I208042b052156599e51da5113a8a350cacdbebda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4626238
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Yves Arrouye <drcrash@chromium.org>