cryptohome: Get rid of UserExists from keyset management
It's a bit weird to have a few places in the code using this that aren't
really focused on vault keysets anymore. This isn't a high priority
change but it's something I came across while doing another change that
I ended up backing out, but this seems worth keeping.

The actual implementation is simple enough that I think just inlining
the user->path->exists lookup is better than extracting this to a
standalone helper function. But we could still do that if we think it
makes more sense.

BUG=b:281878872
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I4ca58964e80829958f716c49e47d06471d25039f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4630095
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>