arc-setup: Suppress error when parent of stale data dir does not exist
Currently, OnRemoveStaleData() emits an error message when the parent of
a stale android data directory does not exist.
However, such a state is expected especially for the stale ODEX
directory of ARCVM; for example, the directory as well as its parent
does not exist just after successfully finishing arc-remove-data.
This CL adds a check for the existence of the parent of each stale
android data directory in OnRemoveStaleData(). When the parent does not
exist, the removal is skipped without emitting any error message.

BUG=b:288231102
TEST=FEATURES=test emerge-${BOARD} chromeos-base/arc-setup
TEST=.../android-data/data/vendor/arc/old_arc_executables_pre_ota and
TEST=.../android-data-old are removed by the following command:
TEST=start arc-stale-directory-remover CHROMEOS_USER=<user>

Change-Id: I27b85eef2b75f8f85c33ba1268303b71091d8271
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4629095
Reviewed-by: Yury Khmel <khmel@google.com>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>