power: Adapt hibernate to new configuration method
Because of limitations with feature_library which are documented in
the doc linked in b/287032879 we have to change how we will manage the
hibernate experiment. Effectively, feature_library cannot view features
managed by chrome and vice versa. This means we had to restructure things
so that Chrome manages the hibernate experiment and writes out the
feature param to a tmpfs file.

This CL adapts the existing code to that new proposal and adds new tests
and modifies exist test accordingly.

BUG=b:287032879
TEST=New unit tests

Change-Id: Ib45525b79378086612f80f2a765655e407d3bf0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4651888
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Brian Geffon <bgeffon@chromium.org>