cryptohome: Make a verify-based GetSupportedIntents
This unifies the persistent vs ephemeral factor-with-status lookups
more. The idea is that when we have dynamic intents available we can put
the logic for checking if an intent is enabled directly into
GetSupportedIntents (or maybe a new GetEnabledIntents wrapper) and this
is easier if we can reuse the same logic in the ephemeral paths.

BUG=b:281878872
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I63432ba5a0b65f0497352425e907fb1f3d3ea79e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4631925
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>