thinpool_migrator: Add recovery from reboot/crashes
Add VPD-backed state for the migrator: this includes the state
of the migration and the number of tries available.

The migrator will attempt to complete the migration in case of
a sudden reboot/power reset until there are tries available.

Once all tries have been exhausted, the migrator will revert the
migration entirely.

BUG=b:232006113
TEST=unittests

Change-Id: I0378fad75fd76324ac769a9d4afaef0a34d6b72e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4662489
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>