power: Add support for charge_limit_enabled to check_powerd_config
This adds support for checking the charge_limit_enabled pref via
check_powerd_config with the --charge_limit_enabled flag. This is needed
for tests to check if charge limit is enabled, so that tests that
require it to be disabled may override the existing setting.

BUG=b:284532839
TEST=emerge-atlas power_manager

Change-Id: Ifc6c541d5547a60a08709b9a278d44f77de9729f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4635104
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>