missive: Implement per-priority storage flag
The flag is represented as a comma-separated string listing all
priorities that are to be kept in legacy (single-generation action)
mode.

Also modified StorageModule to no longer be able to shut down and
restart, since this is not going to happen anymore: instead we will
just reconfigure it, switching each priority from/to multi-generation
action on the fly.

BUG=b:278121325, b:281715983
TEST=FEATURES=test emerge-${BOARD} missive
TEST=Tested manually on VM and DUT

Change-Id: I914e9753a97c244dfcb1238c709ccea9e606ba3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4638124
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>